added change position & click event

+++ js/alert.js
@@ -5,22 +5,50 @@ class Alert {
     this.alert = alert;
   }
 
-  create(message, style, duration = null) {
+  create(options) {
+    /* Create message & type */
     let p = document.createElement('p');
-    p.innerText = message;
-    this.alert.classList = style;
+    p.innerText = options.message;
+    this.alert.classList = options.type;
     this.alert.appendChild(p);
     document.body.appendChild(this.alert);
 
-    this.alert.addEventListener('click', () => {
-      this.delete();
-    });
+    /* Alert position */
+    if (!options.position) {
+      options.position = {
+        top: '20px',
+        right: '20px',
+      };
+    }
+    if (options.position) {
+      options.position.top
+        ? (this.alert.style.top = options.position.top)
+        : null;
+      options.position.bottom
+        ? (this.alert.style.bottom = options.position.bottom)
+        : null;
+      options.position.left
+        ? (this.alert.style.left = options.position.left)
+        : null;
+      options.position.right
+        ? (this.alert.style.right = options.position.right)
+        : null;
+    }
 
-    if (duration != null) {
+    /* Auto remove */
+    if (options.duration) {
       setTimeout(() => {
         this.delete();
-      }, duration);
+      }, options.duration);
     }
+    /* Click remove */
+    this.alert.addEventListener('click', () => {
+      if (!options.click) {
+        this.delete();
+        return false;
+      }
+      options.click();
+    });
   }
 
   delete() {
@@ -36,5 +64,15 @@ const alert = new Alert();
 
 setTimeout(() => {
   // Create Alert
-  alert.create('How quickly daft jumping zebras vex. Alert Link.', 'error');
+  alert.create({
+    message: 'www.google.com adresine git.',
+    type: 'info',
+    position: {
+      top: '40px',
+      left: '50px',
+    },
+    click: () => {
+      window.location.href = 'http://www.google.com';
+    },
+  });
 }, 2500);
